Refactor code editor container (#1999)
