--- src/main/webapp/app/exam/participate/exercises/programming/code-editor/exam-code-editor-student-container.component.html
@@ -1,67 +1,0 @@
-<jhi-code-editor-grid #grid [exerciseTitle]="exercise?.title" [exerciseMaxScore]="exercise?.maxScore">
-    <div class="d-flex align-items-center ml-auto" editorNavbar>
-        <jhi-updating-result
-            *ngIf="participation"
-            [exercise]="exercise"
-            [showUngradedResults]="true"
-            [showGradedBadge]="true"
-            [participation]="participation"
-            [short]="true"
-            [personalParticipation]="true"
-            class="mr-2"
-        ></jhi-updating-result>
-        <jhi-programming-exercise-student-trigger-build-button
-            class="mr-2"
-            [exercise]="exercise"
-            [participation]="participation"
-            [btnSize]="ButtonSize.MEDIUM"
-        ></jhi-programming-exercise-student-trigger-build-button>
-        <jhi-code-editor-actions
-            [disableActions]="repositoryIsLocked"
-            [unsavedFiles]="unsavedFiles"
-            [(editorState)]="editorState"
-            [(commitState)]="commitState"
-            (onSavedFiles)="onSavedFiles($event)"
-            (onRefreshFiles)="onRefreshFiles()"
-            (onError)="onError($event)"
-            (commitStateChange)="onCommitStateChange($event)"
-        ></jhi-code-editor-actions>
-    </div>
-    <jhi-code-editor-file-browser
-        editorSidebarLeft
-        [disableActions]="repositoryIsLocked"
-        [exerciseTitle]="exercise?.title"
-        [unsavedFiles]="unsavedFiles | keys"
-        [errorFiles]="buildLogErrors ? (buildLogErrors.errors | keys) : []"
-        [editorState]="editorState"
-        [(selectedFile)]="selectedFile"
-        [(commitState)]="commitState"
-        (onFileChange)="onFileChange($event)"
-        (onError)="onError($event)"
-        (onToggleCollapse)="onToggleCollapse($event)"
-    ></jhi-code-editor-file-browser>
-    <jhi-code-editor-ace
-        editorCenter
-        [selectedFile]="selectedFile"
-        [fileChange]="fileChange"
-        [commitState]="commitState"
-        [editorState]="editorState"
-        [(buildLogErrors)]="buildLogErrors"
-        (onFileContentChange)="onFileContentChange($event)"
-        (onError)="onError($event)"
-    ></jhi-code-editor-ace>
-    <jhi-code-editor-instructions
-        editorSidebarRight
-        [participation]="participation"
-        [exercise]="exercise"
-        (onToggleCollapse)="onToggleCollapse($event)"
-        (onError)="onError($event)"
-    ></jhi-code-editor-instructions>
-    <jhi-code-editor-build-output
-        editorBottomArea
-        [participation]="participation"
-        [(buildLogErrors)]="buildLogErrors"
-        (onError)="onError($event)"
-        (onToggleCollapse)="onToggleCollapse($event)"
-    ></jhi-code-editor-build-output>
-</jhi-code-editor-grid>
