+++ src/main/webapp/app/exercises/programming/manage/code-editor/code-editor-instructor-container.component.ts
@@ -4,32 +4,22 @@ import { Location } from '@angular/common';
 import { TranslateService } from '@ngx-translate/core';
 import { AlertService } from 'app/core/alert/alert.service';
 import { UpdatingResultComponent } from 'app/exercises/shared/result/updating-result.component';
-import { CodeEditorAceComponent } from 'app/exercises/programming/shared/code-editor/ace/code-editor-ace.component';
-import { CodeEditorActionsComponent } from 'app/exercises/programming/shared/code-editor/actions/code-editor-actions.component';
-import { CodeEditorBuildOutputComponent } from 'app/exercises/programming/shared/code-editor/build-output/code-editor-build-output.component';
-import { CodeEditorFileBrowserComponent } from 'app/exercises/programming/shared/code-editor/file-browser/code-editor-file-browser.component';
-import { CodeEditorInstructionsComponent } from 'app/exercises/programming/shared/code-editor/instructions/code-editor-instructions.component';
 import { CodeEditorInstructorBaseContainerComponent } from 'app/exercises/programming/manage/code-editor/code-editor-instructor-base-container.component';
-import { CodeEditorFileService } from 'app/exercises/programming/shared/code-editor/service/code-editor-file.service';
-import { CodeEditorSessionService } from 'app/exercises/programming/shared/code-editor/service/code-editor-session.service';
 import { CourseExerciseService } from 'app/course/manage/course-management.service';
 import { DomainService } from 'app/exercises/programming/shared/code-editor/service/code-editor-domain.service';
 import { ProgrammingExerciseParticipationService } from 'app/exercises/programming/manage/services/programming-exercise-participation.service';
 import { ProgrammingExerciseService } from 'app/exercises/programming/manage/services/programming-exercise.service';
 import { ParticipationService } from 'app/exercises/shared/participation/participation.service';
 import { ExerciseHintService } from 'app/exercises/shared/exercise-hint/manage/exercise-hint.service';
+import { ProgrammingExerciseEditableInstructionComponent } from 'app/exercises/programming/manage/instructions-editor/programming-exercise-editable-instruction.component';
 
 @Component({
     selector: 'jhi-code-editor-instructor',
     templateUrl: './code-editor-instructor-container.component.html',
 })
 export class CodeEditorInstructorContainerComponent extends CodeEditorInstructorBaseContainerComponent {
-    @ViewChild(CodeEditorFileBrowserComponent, { static: false }) fileBrowser: CodeEditorFileBrowserComponent;
-    @ViewChild(CodeEditorActionsComponent, { static: false }) actions: CodeEditorActionsComponent;
-    @ViewChild(CodeEditorBuildOutputComponent, { static: false }) buildOutput: CodeEditorBuildOutputComponent;
-    @ViewChild(CodeEditorInstructionsComponent, { static: false }) instructions: CodeEditorInstructionsComponent;
-    @ViewChild(CodeEditorAceComponent, { static: false }) aceEditor: CodeEditorAceComponent;
     @ViewChild(UpdatingResultComponent, { static: false }) resultComp: UpdatingResultComponent;
+    @ViewChild(ProgrammingExerciseEditableInstructionComponent, { static: false }) editableInstructions: ProgrammingExerciseEditableInstructionComponent;
 
     constructor(
         router: Router,
@@ -43,8 +33,6 @@ export class CodeEditorInstructorContainerComponent extends CodeEditorInstructor
         translateService: TranslateService,
         route: ActivatedRoute,
         jhiAlertService: AlertService,
-        sessionService: CodeEditorSessionService,
-        fileService: CodeEditorFileService,
     ) {
         super(
             router,
@@ -55,11 +43,8 @@ export class CodeEditorInstructorContainerComponent extends CodeEditorInstructor
             exerciseHintService,
             location,
             participationService,
-            translateService,
             route,
             jhiAlertService,
-            sessionService,
-            fileService,
         );
     }
 
@@ -89,5 +74,11 @@ export class CodeEditorInstructorContainerComponent extends CodeEditorInstructor
      */
     selectTestRepository() {
         this.router.navigate(['..', 'test'], { relativeTo: this.route });
+    }
+
+    onResizeEditorInstructions() {
+        if (this.editableInstructions.markdownEditor && this.editableInstructions.markdownEditor.aceEditorContainer) {
+            this.editableInstructions.markdownEditor.aceEditorContainer.getEditor().resize();
+        }
     }
 }
