+++ src/main/webapp/app/exercises/programming/participate/code-editor-student-container.component.ts
@@ -3,44 +3,32 @@ import { Observable, of } from 'rxjs';
 import { Subscription } from 'rxjs/Subscription';
 import { catchError, flatMap, map, switchMap, tap } from 'rxjs/operators';
 import * as moment from 'moment';
-import { ParticipationService } from 'app/exercises/shared/participation/participation.service';
-import { TranslateService } from '@ngx-translate/core';
-import { ActivatedRoute } from '@angular/router';
-import { AlertService } from 'app/core/alert/alert.service';
 import { ProgrammingExerciseParticipationService } from 'app/exercises/programming/manage/services/programming-exercise-participation.service';
 import { GuidedTourService } from 'app/guided-tour/guided-tour.service';
 import { codeEditorTour } from 'app/guided-tour/tours/code-editor-tour';
-import { CodeEditorBuildOutputComponent } from 'app/exercises/programming/shared/code-editor/build-output/code-editor-build-output.component';
 import { ButtonSize } from 'app/shared/components/button.component';
-import { CodeEditorSessionService } from 'app/exercises/programming/shared/code-editor/service/code-editor-session.service';
 import { ResultService } from 'app/exercises/shared/result/result.service';
 import { DomainService } from 'app/exercises/programming/shared/code-editor/service/code-editor-domain.service';
-import { CodeEditorFileService } from 'app/exercises/programming/shared/code-editor/service/code-editor-file.service';
-import { CodeEditorActionsComponent } from 'app/exercises/programming/shared/code-editor/actions/code-editor-actions.component';
-import { CodeEditorAceComponent } from 'app/exercises/programming/shared/code-editor/ace/code-editor-ace.component';
 import { ExerciseType } from 'app/entities/exercise.model';
 import { AssessmentType } from 'app/entities/assessment-type.model';
 import { StudentParticipation } from 'app/entities/participation/student-participation.model';
 import { Result } from 'app/entities/result.model';
-import { CodeEditorContainerComponent } from 'app/exercises/programming/shared/code-editor/code-editor-mode-container.component';
 import { Feedback } from 'app/entities/feedback.model';
-import { CodeEditorInstructionsComponent } from 'app/exercises/programming/shared/code-editor/instructions/code-editor-instructions.component';
-import { CodeEditorFileBrowserComponent } from 'app/exercises/programming/shared/code-editor/file-browser/code-editor-file-browser.component';
 import { ProgrammingExercise } from 'app/entities/programming-exercise.model';
 import { DomainType } from 'app/exercises/programming/shared/code-editor/model/code-editor.model';
 import { ExerciseHint } from 'app/entities/exercise-hint.model';
 import { ExerciseHintService } from 'app/exercises/shared/exercise-hint/manage/exercise-hint.service';
+import { ActivatedRoute } from '@angular/router';
+import { CodeEditorContainerComponent } from 'app/exercises/programming/shared/code-editor/container/code-editor-container.component';
+import { ProgrammingExerciseInstructionComponent } from 'app/exercises/programming/shared/instructions-render/programming-exercise-instruction.component';
+import { ComponentCanDeactivate } from 'app/shared/guard/can-deactivate.model';
 
 @Component({
     selector: 'jhi-code-editor-student',
     templateUrl: './code-editor-student-container.component.html',
 })
-export class CodeEditorStudentContainerComponent extends CodeEditorContainerComponent implements OnInit, OnDestroy {
-    @ViewChild(CodeEditorFileBrowserComponent, { static: false }) fileBrowser: CodeEditorFileBrowserComponent;
-    @ViewChild(CodeEditorActionsComponent, { static: false }) actions: CodeEditorActionsComponent;
-    @ViewChild(CodeEditorBuildOutputComponent, { static: false }) buildOutput: CodeEditorBuildOutputComponent;
-    @ViewChild(CodeEditorInstructionsComponent, { static: false }) instructions: CodeEditorInstructionsComponent;
-    @ViewChild(CodeEditorAceComponent, { static: false }) aceEditor: CodeEditorAceComponent;
+export class CodeEditorStudentContainerComponent implements OnInit, OnDestroy, ComponentCanDeactivate {
+    @ViewChild(CodeEditorContainerComponent, { static: false }) codeEditorContainer: CodeEditorContainerComponent;
 
     ButtonSize = ButtonSize;
     PROGRAMMING = ExerciseType.PROGRAMMING;
@@ -53,6 +41,7 @@ export class CodeEditorStudentContainerComponent extends CodeEditorContainerComp
     loadingParticipation = false;
     participationCouldNotBeFetched = false;
     repositoryIsLocked = false;
+    showEditorInstructions = true;
 
     constructor(
         private resultService: ResultService,
@@ -60,15 +49,8 @@ export class CodeEditorStudentContainerComponent extends CodeEditorContainerComp
         private programmingExerciseParticipationService: ProgrammingExerciseParticipationService,
         private guidedTourService: GuidedTourService,
         private exerciseHintService: ExerciseHintService,
-        participationService: ParticipationService,
-        translateService: TranslateService,
-        route: ActivatedRoute,
-        jhiAlertService: AlertService,
-        sessionService: CodeEditorSessionService,
-        fileService: CodeEditorFileService,
-    ) {
-        super(participationService, translateService, route, jhiAlertService, sessionService, fileService);
-    }
+        private route: ActivatedRoute,
+    ) {}
 
     /**
      * On init set up the route param subscription.
@@ -156,5 +138,9 @@ export class CodeEditorStudentContainerComponent extends CodeEditorContainerComp
      */
     loadResultDetails(result: Result): Observable<Feedback[] | null> {
         return this.resultService.getFeedbackDetailsForResult(result.id).pipe(map((res) => res && res.body));
+    }
+
+    canDeactivate() {
+        return this.codeEditorContainer.canDeactivate();
     }
 }
