+++ src/main/webapp/app/exercises/programming/participate/programming-participation.module.ts
@@ -6,6 +6,7 @@ import { ArtemisSharedModule } from 'app/shared/shared.module';
 import { ArtemisProgrammingExerciseActionsModule } from 'app/exercises/programming/shared/actions/programming-exercise-actions.module';
 import { ArtemisResultModule } from 'app/exercises/shared/result/result.module';
 import { ArtemisExerciseHintParticipationModule } from 'app/exercises/shared/exercise-hint/participate/exercise-hint-participation.module';
+import { ArtemisProgrammingExerciseInstructionsRenderModule } from 'app/exercises/programming/shared/instructions-render/programming-exercise-instructions-render.module';
 
 @NgModule({
     imports: [
@@ -15,6 +16,7 @@ import { ArtemisExerciseHintParticipationModule } from 'app/exercises/shared/exe
         ArtemisProgrammingExerciseActionsModule,
         ArtemisExerciseHintParticipationModule,
         ArtemisResultModule,
+        ArtemisProgrammingExerciseInstructionsRenderModule,
     ],
     declarations: [CodeEditorStudentContainerComponent],
 })
