+++ src/main/webapp/app/exercises/programming/shared/code-editor/code-editor.module.ts
@@ -19,6 +19,7 @@ import { CodeEditorFileBrowserComponent } from 'app/exercises/programming/shared
 import { CodeEditorStatusComponent } from 'app/exercises/programming/shared/code-editor/status/code-editor-status.component';
 import { FeatureToggleModule } from 'app/shared/feature-toggle/feature-toggle.module';
 import { CodeEditorConfirmRefreshModalComponent } from 'app/exercises/programming/shared/code-editor/actions/code-editor-confirm-refresh-modal.component';
+import { CodeEditorContainerComponent } from 'app/exercises/programming/shared/code-editor/container/code-editor-container.component';
 
 @NgModule({
     imports: [AceEditorModule, MomentModule, ArtemisSharedModule, FeatureToggleModule, TreeviewModule.forRoot(), ArtemisProgrammingExerciseInstructionsEditorModule],
@@ -37,6 +38,7 @@ import { CodeEditorConfirmRefreshModalComponent } from 'app/exercises/programmin
         CodeEditorActionsComponent,
         CodeEditorResolveConflictModalComponent,
         CodeEditorConfirmRefreshModalComponent,
+        CodeEditorContainerComponent,
     ],
     exports: [
         CodeEditorGridComponent,
@@ -46,6 +48,7 @@ import { CodeEditorConfirmRefreshModalComponent } from 'app/exercises/programmin
         CodeEditorActionsComponent,
         CodeEditorInstructionsComponent,
         CodeEditorBuildOutputComponent,
+        CodeEditorContainerComponent,
     ],
     entryComponents: [CodeEditorFileBrowserDeleteComponent, CodeEditorResolveConflictModalComponent, CodeEditorConfirmRefreshModalComponent],
     providers: [],
