+++ src/main/webapp/app/exercises/programming/shared/code-editor/container/code-editor-container.component.html
@@ -1,0 +1,53 @@
+<jhi-code-editor-grid #grid (onResize)="onGridResize($event)">
+    <div editorNavbar class="d-flex align-items-center">
+        <h4><ng-content select="[editorTitle]"></ng-content></h4>
+        <ng-content select="[editorTitleActions]"></ng-content>
+        <div class="d-flex align-items-center ml-auto">
+            <ng-content select="[editorToolbar]"></ng-content>
+            <jhi-code-editor-actions
+                [buildable]="buildable"
+                [disableActions]="!editable"
+                [unsavedFiles]="unsavedFiles"
+                [(editorState)]="editorState"
+                [(commitState)]="commitState"
+                (onSavedFiles)="onSavedFiles($event)"
+                (onRefreshFiles)="onRefreshFiles()"
+                (commitStateChange)="onCommitStateChange.emit($event)"
+                (onError)="onError($event)"
+            ></jhi-code-editor-actions>
+        </div>
+    </div>
+    <jhi-code-editor-file-browser
+        editorSidebarLeft
+        [disableActions]="!editable"
+        [unsavedFiles]="unsavedFiles | keys"
+        [errorFiles]="buildLogErrors ? (buildLogErrors.errors | keys) : []"
+        [editorState]="editorState"
+        [(selectedFile)]="selectedFile"
+        [(commitState)]="commitState"
+        (onFileChange)="onFileChange($event)"
+        (onError)="onError($event)"
+        (onToggleCollapse)="onToggleCollapse($event)"
+    ></jhi-code-editor-file-browser>
+    <jhi-code-editor-ace
+        editorCenter
+        [selectedFile]="selectedFile"
+        [fileChange]="fileChange"
+        [commitState]="commitState"
+        [editorState]="editorState"
+        [(buildLogErrors)]="buildLogErrors"
+        (onFileContentChange)="onFileContentChange($event)"
+        (onError)="onError($event)"
+    ></jhi-code-editor-ace>
+    <jhi-code-editor-instructions *ngIf="showEditorInstructions" editorSidebarRight (onToggleCollapse)="onToggleCollapse($event)" (onError)="onError($event)">
+        <ng-content select="[editorSidebar]"></ng-content>
+    </jhi-code-editor-instructions>
+    <jhi-code-editor-build-output
+        editorBottomArea
+        *ngIf="buildable"
+        [participation]="participation"
+        [(buildLogErrors)]="buildLogErrors"
+        (onError)="onError($event)"
+        (onToggleCollapse)="onToggleCollapse($event)"
+    ></jhi-code-editor-build-output>
+</jhi-code-editor-grid>
