+++ src/main/webapp/app/exercises/programming/shared/code-editor/file-browser/code-editor-file-browser.component.ts
@@ -37,8 +37,6 @@ export class CodeEditorFileBrowserComponent implements OnInit, OnChanges, AfterV
     @ViewChild('treeview', { static: false }) treeview: TreeviewComponent;
 
     @Input()
-    exerciseTitle: string;
-    @Input()
     get selectedFile(): string | undefined {
         return this.selectedFileValue;
     }
