+++ src/main/webapp/app/exercises/programming/shared/code-editor/instructions/code-editor-instructions.component.html
@@ -1,4 +1,4 @@
-<div *ngIf="!noInstructionsAvailable" id="cardInstructions" class="instructions-wrapper card border border-lightgrey">
+<div id="cardInstructions" class="instructions-wrapper card border border-lightgrey">
     <div class="card-header d-flex justify-content-center text-white bg-primary" (click)="toggleEditorCollapse($event)">
         <h3 class="card-title">
             <fa-icon [icon]="['far', 'list-alt']"></fa-icon>
@@ -6,26 +6,7 @@
         </h3>
         <fa-icon class="ml-auto" [icon]="'chevron-right'"></fa-icon>
     </div>
-    <jhi-programming-exercise-instructions
-        *ngIf="!editable; else markdownEditorInstructions"
-        (onNoInstructionsAvailable)="onNoInstructionsAvailable()"
-        [exercise]="exercise"
-        [participation]="participation"
-        [personalParticipation]="true"
-        [exerciseHints]="exercise.exerciseHints"
-        class="instructions-wrapper__content card-body p-0"
-    >
-    </jhi-programming-exercise-instructions>
-    <ng-template #markdownEditorInstructions>
-        <jhi-programming-exercise-editable-instructions
-            [(exercise)]="exercise"
-            [(participation)]="participation"
-            [templateParticipation]="templateParticipation"
-            [enableResize]="false"
-            class="instructions-wrapper__content card-body p-0"
-        >
-        </jhi-programming-exercise-editable-instructions>
-    </ng-template>
+    <ng-content></ng-content>
     <!-- Required for resizing; don't remove empty span -->
     <div class="rg-left"><span></span></div>
 </div>
