+++ src/main/webapp/app/exercises/programming/shared/code-editor/layout/code-editor-grid.component.html
@@ -1,12 +1,7 @@
 <div class="editor-wrapper">
     <!-- Top bar -->
     <nav class="editor-statusbar">
-        <!-- Exercise title -->
-        <h4>
-            <span>{{ exerciseTitle }}</span>
-            <span *ngIf="exerciseMaxScore">&nbsp;[{{ exerciseMaxScore }} {{ 'artemisApp.examParticipation.points' | translate }}]</span>
-        </h4>
-        <ng-content select="[editorNavbar]"> </ng-content>
+        <ng-content select="[editorNavbar]"></ng-content>
     </nav>
 
     <div class="editor-main">
